fix(cli): split user-input-stream into builder and runner to satisfy one-class-per-file
