+++ apps/cli/src/utils/user-input-stream.ts
@@ -1,5 +1,6 @@
 import readline from 'node:readline/promises';
 import type { Readable, Writable } from 'node:stream';
+import { UserInputStreamBuilder } from './user-input-stream.builder';
 
 export interface UserInputStreamOptions {
   /** Prompt shown to the user for each input. Defaults to '> '. */
@@ -14,53 +15,15 @@ export interface UserInputStreamOptions {
 
 export type InputCallback = (match: RegExpMatchArray) => void | Promise<void>;
 
-interface Matcher {
+export interface Matcher {
   pattern: RegExp;
   callback: InputCallback;
-}
-
-/**
- * Builder for UserInputStream. Register patterns using {@link on} and then
- * call {@link build} to create the stream runner.
- */
-export class UserInputStreamBuilder {
-  private matchers: Matcher[] = [];
-  private quitPatterns: RegExp[];
-
-  constructor(private readonly options: UserInputStreamOptions = {}) {
-    this.quitPatterns = options.quitPatterns?.slice() ?? [/^(quit|exit)$/i];
-  }
-
-  /**
-   * Register a pattern and its handler.
-   * @param pattern Regular expression to match user input
-   * @param callback Invoked with the RegExp match result on match
-   */
-  on(pattern: RegExp, callback: InputCallback): this {
-    this.matchers.push({ pattern, callback });
-    return this;
-  }
-
-  /**
-   * Register an additional command that will exit the input loop when entered.
-   * @param pattern Command string or RegExp to match
-   */
-  quit(pattern: string | RegExp): this {
-    const regex = typeof pattern === 'string' ? new RegExp(`^${pattern}$`, 'i') : pattern;
-    this.quitPatterns.push(regex);
-    return this;
-  }
-
-  build(): UserInputStream {
-    return new UserInputStream(this.matchers, { ...this.options, quitPatterns: this.quitPatterns });
-  }
-}
-
-/**
+} /**
  * User input stream processor. Use {@link UserInputStreamBuilder} to create an
  * instance. Call {@link run} to start handling input. The promise resolves when
  * the user types `quit` or `exit`, and rejects if a handler throws.
  */
+
 export class UserInputStream {
   private readonly prompt: string;
   private readonly input: Readable;
